requirements.txt
# Core
yfinance
pandas
numpy
python-dotenv

# LangChain / LangGraph
langchain
langchain-core
langchain-groq
langgraph

# RAG 
faiss-cpu
sentence-transformers
feedparser

# LLM inference 
groq

# News agent (FinBERT)
transformers
openpyxl
feedparser
rapidfuzz
requests

# Technical Agent 
growwapi

# Fundamental Agent
requests
beautifulsoup4